feat: process german variations
Created basic processing for variations. Next up testing edge cases.
