+++ apps/db_management/eslintrc.js
@@ -4,67 +4,20 @@ module.exports = {
     node: true,
   },
   extends: [
+    'airbnb-base',
     'plugin:@typescript-eslint/recommended',
     'plugin:@typescript-eslint/recommended-requiring-type-checking',
   ],
-  root: true,
   ignorePatterns: ['/*.*', '.eslintrc.js'],
-  overrides: [
-    {
-      files: ['*.ts', '*.tsx', '*.js', '*.jsx'],
-      rules: {
-        '@nrwl/nx/enforce-module-boundaries': [
-          'error',
-          {
-            enforceBuildableLibDependency: true,
-            allow: [],
-            depConstraints: [
-              {
-                sourceTag: '*',
-                onlyDependOnLibsWithTags: ['*'],
-              },
-            ],
-          },
-        ],
-      },
-    },
-    {
-      files: ['*.ts', '*.tsx'],
-      extends: [
-        'plugin:@nrwl/nx/typescript',
-        'airbnb',
-        'airbnb-typescript',
-        'plugin:prettier/recommended',
-        'plugin:react/jsx-runtime',
-      ],
-      rules: { 'no-console': 1 },
-      parserOptions: {
-        project: './tsconfig.base.json',
-      },
-    },
-    {
-      files: ['*.js', '*.jsx'],
-      extends: ['plugin:@nrwl/nx/javascript'],
-      rules: {},
-    },
-    {
-      files: ['*.spec.ts', '*.spec.tsx', '*.spec.js', '*.spec.jsx'],
-      env: {
-        jest: true,
-      },
-      rules: {},
-    },
-  ],
   parser: '@typescript-eslint/parser',
   parserOptions: {
     ecmaVersion: 13,
-    project: 'tsconfig.base.json',
     sourceType: 'module',
+    project: 'tsconfig.json',
     tsconfigRootDir: __dirname,
   },
-  plugins: ['@nrwl/nx'],
+  plugins: ['@typescript-eslint', 'import'],
   rules: {
-    'no-console': 1,
     'no-restricted-syntax': 0,
     'import/prefer-default-export': 'off',
     'import/extensions': [
