+++ apps/db_management/src/models/german/utilities/modifiedStem.ts
@@ -2,29 +2,33 @@ import { RegExpGroups } from 'global-types';
 
 export const modifiedStem = (stem: string, irregularStem: string): string => {
   const irregularStemRegex =
-    /\b(?<firstConst>[bcdfghjklmnpqrstvwxyzß]*)(?<vowelGroup>[aeiouäöü]*)(?<secondConst>[bcdfghjklmnpqrstvwxyzß]*)\b/;
+    /(?<firstConst>[bcdfghjklmnpqrstvwxyzß]*)(?<vowelGroup>[aeiouäöü]*)(?<secondConst>[bcdfghjklmnpqrstvwxyzß]*)/;
 
   const irregularStemResult: RegExpGroups<
     'firstConst' | 'vowelGroup' | 'secondConst'
   > = irregularStemRegex.exec(irregularStem);
 
-  const {
-    firstConst: irregularStemFirstConst,
-    vowelGroup: irregularStemVowelGroup,
-    secondConst: irregularStemSecondGroup,
-  } = irregularStemResult?.groups!;
-
   const regularStemResult: RegExpGroups<
     'firstConst' | 'vowelGroup' | 'secondConst'
   > = irregularStemRegex.exec(stem);
 
-  const { firstConst, vowelGroup, secondConst } = regularStemResult?.groups!;
+  if (regularStemResult?.groups && irregularStemResult?.groups) {
+    const {
+      firstConst: irregularStemFirstConst,
+      vowelGroup: irregularStemVowelGroup,
+      secondConst: irregularStemSecondGroup,
+    } = irregularStemResult?.groups!;
 
-  if (irregularStemFirstConst && !irregularStemVowelGroup) {
-    return `${firstConst || ''}${vowelGroup}${irregularStemFirstConst}`;
+    const { firstConst, vowelGroup, secondConst } = regularStemResult?.groups!;
+
+    if (irregularStemFirstConst && !irregularStemVowelGroup) {
+      return `${firstConst || ''}${vowelGroup}${irregularStemFirstConst}`;
+    }
+
+    return `${irregularStemFirstConst || firstConst || ''}${
+      irregularStemVowelGroup || vowelGroup || ''
+    }${irregularStemSecondGroup || secondConst || ''}`;
   }
 
-  return `${irregularStemFirstConst || firstConst || ''}${
-    irregularStemVowelGroup || vowelGroup || ''
-  }${irregularStemSecondGroup || secondConst || ''}`;
+  throw Error(`${stem} did not pass regex.`);
 };
